pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pharma-drf-service"
version = "0.1.0"
description = "Production-style DRF service with JWT, RBAC, caching, and audit logs"
authors = [
    {name = "Mohamed Alcafory", email = "mohamed.alcafory.456@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "Django>=5.0.0,<5.1.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-redis>=5.4.0",
    "django-filter>=23.0",
    "drf-spectacular>=0.27.0",
    "psycopg2-binary>=2.9.0",
    "python-decouple>=3.8",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
    "django-stubs>=4.2.0",
    "djangorestframework-stubs>=3.14.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "UP", "S", "B", "A", "C4", "DJ"]
ignore = ["E501", "S101"]

[tool.ruff.per-file-ignores]
"*/tests/*" = ["S101", "S106"]
"*/migrations/*" = ["E501", "DJ"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
plugins = ["mypy_django_plugin.main"]
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--cov=. --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]

[tool.coverage.run]
source = "."
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/tests/*",
    "manage.py",
    "config/*",
    "*/wsgi.py",
    "*/__pycache__/*"
]